Split a big test into multiple ones
