+++ airbyte-workers/src/test/java/io/airbyte/workers/temporal/scheduling/connection/manager/BaseConnectionManagerWorkflowTest.java
@@ -1,0 +1,237 @@
+package io.airbyte.workers.temporal.scheduling.connection.manager;
+
+import io.airbyte.config.ConnectorJobOutput;
+import io.airbyte.config.ConnectorJobOutput.OutputType;
+import io.airbyte.config.FailureReason;
+import io.airbyte.config.FailureReason.FailureOrigin;
+import io.airbyte.config.FailureReason.FailureType;
+import io.airbyte.config.StandardCheckConnectionOutput;
+import io.airbyte.config.StandardCheckConnectionOutput.Status;
+import io.airbyte.config.StandardSyncInput;
+import io.airbyte.scheduler.models.IntegrationLauncherConfig;
+import io.airbyte.scheduler.models.JobRunConfig;
+import io.airbyte.workers.WorkerConstants;
+import io.airbyte.workers.temporal.TemporalJobType;
+import io.airbyte.workers.temporal.check.connection.CheckConnectionActivity;
+import io.airbyte.workers.temporal.scheduling.ConnectionManagerWorkflow;
+import io.airbyte.workers.temporal.scheduling.ConnectionManagerWorkflowImpl;
+import io.airbyte.workers.temporal.scheduling.ConnectionUpdaterInput;
+import io.airbyte.workers.temporal.scheduling.activities.AutoDisableConnectionActivity;
+import io.airbyte.workers.temporal.scheduling.activities.AutoDisableConnectionActivity.AutoDisableConnectionOutput;
+import io.airbyte.workers.temporal.scheduling.activities.ConfigFetchActivity;
+import io.airbyte.workers.temporal.scheduling.activities.ConfigFetchActivity.ScheduleRetrieverOutput;
+import io.airbyte.workers.temporal.scheduling.activities.ConnectionDeletionActivity;
+import io.airbyte.workers.temporal.scheduling.activities.GenerateInputActivity.GeneratedJobInput;
+import io.airbyte.workers.temporal.scheduling.activities.GenerateInputActivity.SyncInputWithAttemptNumber;
+import io.airbyte.workers.temporal.scheduling.activities.GenerateInputActivityImpl;
+import io.airbyte.workers.temporal.scheduling.activities.JobCreationAndStatusUpdateActivity;
+import io.airbyte.workers.temporal.scheduling.activities.JobCreationAndStatusUpdateActivity.AttemptNumberCreationOutput;
+import io.airbyte.workers.temporal.scheduling.activities.JobCreationAndStatusUpdateActivity.AttemptNumberFailureInput;
+import io.airbyte.workers.temporal.scheduling.activities.JobCreationAndStatusUpdateActivity.JobCancelledInputWithAttemptNumber;
+import io.airbyte.workers.temporal.scheduling.activities.JobCreationAndStatusUpdateActivity.JobCreationOutput;
+import io.airbyte.workers.temporal.scheduling.activities.StreamResetActivity;
+import io.airbyte.workers.temporal.scheduling.state.listener.TestStateListener;
+import io.airbyte.workers.temporal.sync.SyncWorkflow;
+import io.temporal.api.enums.v1.WorkflowExecutionStatus;
+import io.temporal.api.filter.v1.WorkflowExecutionFilter;
+import io.temporal.api.workflowservice.v1.ListClosedWorkflowExecutionsRequest;
+import io.temporal.api.workflowservice.v1.ListClosedWorkflowExecutionsResponse;
+import io.temporal.client.WorkflowClient;
+import io.temporal.client.WorkflowOptions;
+import io.temporal.testing.TestWorkflowEnvironment;
+import io.temporal.worker.Worker;
+import java.time.Duration;
+import java.util.stream.Stream;
+import lombok.extern.slf4j.Slf4j;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.mockito.ArgumentMatcher;
+import org.mockito.Mockito;
+
+@Slf4j
+public abstract class BaseConnectionManagerWorkflowTest {
+  protected static final long JOB_ID = 1L;
+  protected static final int ATTEMPT_ID = 1;
+
+  protected static final Duration SCHEDULE_WAIT = Duration.ofMinutes(20L);
+  protected static final String WORKFLOW_ID = "workflow-id";
+
+  protected final ConfigFetchActivity mConfigFetchActivity =
+      Mockito.mock(ConfigFetchActivity.class, Mockito.withSettings().withoutAnnotations());
+  protected final CheckConnectionActivity mCheckConnectionActivity =
+      Mockito.mock(CheckConnectionActivity.class, Mockito.withSettings().withoutAnnotations());
+  protected static final ConnectionDeletionActivity mConnectionDeletionActivity =
+      Mockito.mock(ConnectionDeletionActivity.class, Mockito.withSettings().withoutAnnotations());
+  protected static final GenerateInputActivityImpl mGenerateInputActivityImpl =
+      Mockito.mock(GenerateInputActivityImpl.class, Mockito.withSettings().withoutAnnotations());
+  protected static final JobCreationAndStatusUpdateActivity mJobCreationAndStatusUpdateActivity =
+      Mockito.mock(JobCreationAndStatusUpdateActivity.class, Mockito.withSettings().withoutAnnotations());
+  protected static final AutoDisableConnectionActivity mAutoDisableConnectionActivity =
+      Mockito.mock(AutoDisableConnectionActivity.class, Mockito.withSettings().withoutAnnotations());
+  protected static final StreamResetActivity mStreamResetActivity =
+      Mockito.mock(StreamResetActivity.class, Mockito.withSettings().withoutAnnotations());
+  protected static final String EVENT = "event = ";
+
+  protected TestWorkflowEnvironment testEnv;
+  protected WorkflowClient client;
+  protected ConnectionManagerWorkflow workflow;
+
+  @BeforeEach
+  void setUp() {
+    Mockito.reset(mConfigFetchActivity);
+    Mockito.reset(mCheckConnectionActivity);
+    Mockito.reset(mConnectionDeletionActivity);
+    Mockito.reset(mGenerateInputActivityImpl);
+    Mockito.reset(mJobCreationAndStatusUpdateActivity);
+    Mockito.reset(mAutoDisableConnectionActivity);
+    Mockito.reset(mStreamResetActivity);
+
+    // default is to wait "forever"
+    Mockito.when(mConfigFetchActivity.getTimeToWait(Mockito.any())).thenReturn(new ScheduleRetrieverOutput(
+        Duration.ofDays(100 * 365)));
+
+    Mockito.when(mJobCreationAndStatusUpdateActivity.createNewJob(Mockito.any()))
+        .thenReturn(new JobCreationOutput(
+            1L));
+
+    Mockito.when(mJobCreationAndStatusUpdateActivity.createNewAttemptNumber(Mockito.any()))
+        .thenReturn(new AttemptNumberCreationOutput(
+            1));
+
+    Mockito.when(mGenerateInputActivityImpl.getSyncWorkflowInputWithAttemptNumber(Mockito.any(SyncInputWithAttemptNumber.class)))
+        .thenReturn(
+            new GeneratedJobInput(
+                new JobRunConfig(),
+                new IntegrationLauncherConfig().withDockerImage("some_source"),
+                new IntegrationLauncherConfig(),
+                new StandardSyncInput()));
+
+    Mockito.when(mCheckConnectionActivity.runWithJobOutput(Mockito.any()))
+        .thenReturn(new ConnectorJobOutput().withOutputType(OutputType.CHECK_CONNECTION)
+            .withCheckConnection(new StandardCheckConnectionOutput().withStatus(Status.SUCCEEDED).withMessage("check worked")));
+
+    Mockito.when(mAutoDisableConnectionActivity.autoDisableFailingConnection(Mockito.any()))
+        .thenReturn(new AutoDisableConnectionOutput(false));
+  }
+
+  @AfterEach
+  void tearDown() {
+    testEnv.shutdown();
+    TestStateListener.reset();
+  }
+
+  protected void mockResetJobInput() {
+    Mockito.when(mGenerateInputActivityImpl.getSyncWorkflowInputWithAttemptNumber(Mockito.any(SyncInputWithAttemptNumber.class)))
+        .thenReturn(
+            new GeneratedJobInput(
+                new JobRunConfig(),
+                new IntegrationLauncherConfig().withDockerImage(WorkerConstants.RESET_JOB_SOURCE_DOCKER_IMAGE_STUB),
+                new IntegrationLauncherConfig(),
+                new StandardSyncInput()));
+  }
+
+  protected static void startWorkflowAndWaitUntilReady(final ConnectionManagerWorkflow workflow, final ConnectionUpdaterInput input)
+      throws InterruptedException {
+    WorkflowClient.start(workflow::run, input);
+
+    boolean isReady = false;
+
+    while (!isReady) {
+      try {
+        isReady = workflow.getState() != null;
+      } catch (final Exception e) {
+        log.info("retrying...");
+        Thread.sleep(100);
+      }
+    }
+  }
+
+  protected <T extends SyncWorkflow> void setupSpecificChildWorkflow(final Class<T> mockedSyncedWorkflow) {
+    testEnv = TestWorkflowEnvironment.newInstance();
+
+    final Worker syncWorker = testEnv.newWorker(TemporalJobType.SYNC.name());
+    syncWorker.registerWorkflowImplementationTypes(mockedSyncedWorkflow);
+
+    final Worker managerWorker = testEnv.newWorker(TemporalJobType.CONNECTION_UPDATER.name());
+    managerWorker.registerWorkflowImplementationTypes(ConnectionManagerWorkflowImpl.class);
+    managerWorker.registerActivitiesImplementations(mConfigFetchActivity, mCheckConnectionActivity, mConnectionDeletionActivity,
+        mGenerateInputActivityImpl, mJobCreationAndStatusUpdateActivity, mAutoDisableConnectionActivity, mStreamResetActivity);
+
+    client = testEnv.getWorkflowClient();
+    testEnv.start();
+
+    workflow = client
+        .newWorkflowStub(
+            ConnectionManagerWorkflow.class,
+            WorkflowOptions.newBuilder()
+                .setTaskQueue(TemporalJobType.CONNECTION_UPDATER.name())
+                .setWorkflowId(WORKFLOW_ID)
+                .build());
+  }
+
+  protected void assertWorkflowWasContinuedAsNew() {
+    final ListClosedWorkflowExecutionsRequest request = ListClosedWorkflowExecutionsRequest.newBuilder()
+        .setNamespace(testEnv.getNamespace())
+        .setExecutionFilter(WorkflowExecutionFilter.newBuilder().setWorkflowId(WORKFLOW_ID))
+        .build();
+    final ListClosedWorkflowExecutionsResponse listResponse = testEnv
+        .getWorkflowService()
+        .blockingStub()
+        .listClosedWorkflowExecutions(request);
+    Assertions.assertThat(listResponse.getExecutionsCount()).isGreaterThanOrEqualTo(1);
+    Assertions.assertThat(listResponse.getExecutionsList().get(0).getStatus())
+        .isEqualTo(WorkflowExecutionStatus.WORKFLOW_EXECUTION_STATUS_CONTINUED_AS_NEW);
+  }
+
+  protected class HasFailureFromOrigin implements ArgumentMatcher<AttemptNumberFailureInput> {
+
+    private final FailureOrigin expectedFailureOrigin;
+
+    HasFailureFromOrigin(final FailureOrigin failureOrigin) {
+      this.expectedFailureOrigin = failureOrigin;
+    }
+
+    @Override
+    public boolean matches(final AttemptNumberFailureInput arg) {
+      return arg.getAttemptFailureSummary().getFailures().stream().anyMatch(f -> f.getFailureOrigin().equals(expectedFailureOrigin));
+    }
+
+  }
+
+  protected class HasFailureFromOriginWithType implements ArgumentMatcher<AttemptNumberFailureInput> {
+
+    private final FailureOrigin expectedFailureOrigin;
+    private final FailureType expectedFailureType;
+
+    HasFailureFromOriginWithType(final FailureOrigin failureOrigin, final FailureType failureType) {
+      this.expectedFailureOrigin = failureOrigin;
+      this.expectedFailureType = failureType;
+    }
+
+    @Override
+    public boolean matches(final AttemptNumberFailureInput arg) {
+      final Stream<FailureReason> stream = arg.getAttemptFailureSummary().getFailures().stream();
+      return stream.anyMatch(f -> f.getFailureOrigin().equals(expectedFailureOrigin) && f.getFailureType().equals(expectedFailureType));
+    }
+
+  }
+
+  protected class HasCancellationFailure implements ArgumentMatcher<JobCancelledInputWithAttemptNumber> {
+
+    private final long expectedJobId;
+    private final int expectedAttemptNumber;
+
+    HasCancellationFailure(final long jobId, final int attemptNumber) {
+      this.expectedJobId = jobId;
+      this.expectedAttemptNumber = attemptNumber;
+    }
+
+    @Override
+    public boolean matches(final JobCancelledInputWithAttemptNumber arg) {
+      return arg.getAttemptFailureSummary().getFailures().stream().anyMatch(f -> f.getFailureType().equals(FailureType.MANUAL_CANCELLATION))
+          && arg.getJobId() == expectedJobId && arg.getAttemptNumber() == expectedAttemptNumber;
+    }
+
+  }
+}
